requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-appconf==1.0.2
django-invitations==1.9.2
django-mathfilters==0.4.0
django-select2==6.0.2
django-simple-history==2.0
idna==2.6
oauthlib==2.0.6
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
sqlparse==0.2.4
urllib3==1.22
